requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2

